moved precheck logic from core.py into precheck.py
