+++ test/test_precheck.py
@@ -4,7 +4,7 @@ import logging
 from supy._env import logger_supy
 from copy import deepcopy
 from datetime import datetime
-from supy.data_model.core import (
+from supy.data_model.precheck import (
     precheck_model_physics_params,
     precheck_start_end_date,
     precheck_site_season_adjustments,
